+++ tests/spec/templates/schema.ts
@@ -1,5 +1,13 @@
 /* tslint:disable */
 /* eslint-disable */
+/*
+ * ---------------------------------------------------------------
+ * ## THIS FILE WAS GENERATED VIA SWAGGER-TYPESCRIPT-API        ##
+ * ##                                                           ##
+ * ## AUTHOR: acacode                                           ##
+ * ## SOURCE: https://github.com/acacode/swagger-typescript-api ##
+ * ---------------------------------------------------------------
+ */
 
 export interface Pet {
   id: number;
@@ -112,7 +120,6 @@ class HttpClient<SecurityDataType> {
     });
   };
 }
-
 /**
  * @title Swagger Petstore
  * @version 1.0.0