+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# next release  
+
+BREAKING_CHANGES:  
+- Renamed mustache templates:  
+  - `api.mustache` -> `data-contracts.mustache`   
+  - `client.mustache` -> `http.client.mustache` + `api.mustache`  
+- Split the `client.mustache` template into two parts: `http-client.mustache` and `api.mustache`
+
 # 2.0.0  
 
 Features: